requirements.txt
# LlamaIndex core and integrations
llama-index>=0.10.0
llama-index-llms-google-genai>=0.2.0
llama-index-embeddings-google-genai>=0.2.0

# Microsoft Azure and Power Platform
azure-identity>=1.12.0
PowerPlatform-Dataverse-Client>=0.1.0b3

# HTTP and utilities
requests>=2.28.0
python-dotenv>=1.0.0
